requirements.txt.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
ecdsa==0.19.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-pagination==0.12.34
fonttools==4.56.0
h11==0.14.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
pip==24.3.1
pyasn1==0.6.1
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.1
qrcode==8.0
requests==2.32.3
rsa==4.9
setuptools==63.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
